+++ src/test/java/org/scion/jpan/internal/DNSHelperTest.java
@@ -184,6 +184,33 @@ class DNSHelperTest {
     assertEquals(expected, output);
   }
 
+  @Test
+  void testNAPTROrderPriority() {
+    DNSUtil.installAddress("whoami.akamai.net", new byte[] {1, 2, 3, 4});
+    DNSUtil.installPTR("4.3.2.1.in-addr.arpa.", "my-dhcp-122-133-233-773.inf.hello.test");
+
+    // The DNS cache processes entries in installation order.
+
+    // Bad order
+    DNSUtil.installNAPTR_only("hello.test", "S", "x-sciondiscovery:tcp", "hi21.test.", 2, 10);
+    DNSUtil.installSRV("hi21.test.", "discovery21.test", 20010);
+    DNSUtil.installAddress("discovery21.test", new byte[] {1, 1, 1, 1});
+
+    // Good order, bad preference
+    DNSUtil.installNAPTR_only("hello.test", "S", "x-sciondiscovery:tcp", "hi12.test.", 1, 20);
+    DNSUtil.installSRV("hi12.test.", "discovery12.test", 10020);
+    DNSUtil.installAddress("discovery12.test", new byte[] {1, 1, 1, 1});
+
+    // Good order + good preference -> This is it!
+    DNSUtil.installNAPTR_only("hello.test", "S", "x-sciondiscovery:tcp", "hi11.test.", 1, 10);
+    DNSUtil.installSRV("hi11.test.", "discovery11.test", 10010);
+    DNSUtil.installAddress("discovery11.test", new byte[] {1, 1, 1, 1});
+
+    Lookup.setDefaultSearchPath(Collections.emptyList());
+    InetSocketAddress dsAddress = DNSHelper.searchForDiscoveryService(new MockDNS.MockResolver());
+    assertEquals("1.1.1.1:10010", IPHelper.toString(dsAddress));
+  }
+
   private InetAddress findSubnet(int len) {
     for (InetAddress i : IPHelper.getSubnets()) {
       if (i.getAddress().length == len) {